replace old master with new architecture

+++ src/test/java/eu/freme/common/persistence/TokenRepositoryTest.java
@@ -22,6 +22,7 @@ import eu.freme.common.persistence.dao.TokenDAO;
 import eu.freme.common.persistence.dao.UserDAO;
 import eu.freme.common.persistence.model.Token;
 import eu.freme.common.persistence.model.User;
+
 import org.apache.log4j.Logger;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -73,7 +74,7 @@ public class TokenRepositoryTest {
 		Token token2 = new Token("t2", user);
 		token = tokenDAO.save(token2);
 		logger.info("token count (before delete): " + tokenDAO.count());
-		assertEquals((long) 2, tokenDAO.count());
+		assertEquals(2, tokenDAO.count());
 		tokenDAO.delete(token);
 
 		logger.info("token count (after delete): " + tokenDAO.count());
